robot/requirements.txt
aiohttp>=3.7.4
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
discord.py==1.6.0
DiscordUtils==1.3.4
httplib2==0.19.0
idna==2.10
multidict==4.7.6
nltk == 3.6.7
oauthlib==3.1.0
prettyprint==0.1.5
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.3
pyparsing==2.4.7
python-dotenv==0.14.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa>=4.7
six==1.15.0
soupsieve==2.0.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3>=1.26.5
yarl==1.6.0
